+++ app/src/main/java/com/example/mobileappstrusted/dataclass/WavBlock.kt
@@ -3,5 +3,6 @@ package com.example.mobileappstrusted.dataclass
 data class WavBlock(
     val originalIndex: Int,
     var currentIndex: Int,
-    val data: ByteArray
+    val data: ByteArray,
+    val isDeleted: Boolean
 )
